重构选择生效路由

+++ lib/port.py
@@ -5,11 +5,8 @@ from tools.telnetClient import telnetClient
 import re
 import logging
 
-import logging
-
 logging.basicConfig(level=logging.DEBUG,
                     format='%(asctime)s - %(filename)s[line:%(lineno)d func:%(funcName)s] - %(levelname)s: %(message)s')
-
 
 class Port:
     def __init__(self, **kwargs):
@@ -21,8 +18,9 @@ class Port:
         self.__type = 0
         self.__isUp = False
         self.__pattern = ""
+        self.__isActivate = False
         if 'conf' in kwargs:
-            self.__initByFile__(kwargs['conf'])
+            self.changePort(kwargs['conf'])
 
     def __configPort__(self):
         try:
@@ -31,12 +29,21 @@ class Port:
             logging.info("配置port失败." + self._get_msg())
             # print("连接失败")
 
+    #激活的同时，进行配置
+    def activate(self):
+        self.__isActivate = True
+        self.__configPort__()
+
+    def deActivate(self):
+        self.deleteConf()
+        self.__isActivate = False
+
     def deleteConf(self):
-        try:
-            telnetClient.delete_port_conf(self.__name, self.__ip, self.__mask)
-        except:
-            logging.info("删除port配置失败." + self._get_msg())
-            # print("连接失败")
+        if self.__isActivate:
+            try:
+                telnetClient.delete_port_conf(self.__name, self.__ip, self.__mask)
+            except:
+                logging.info("删除port配置失败." + self._get_msg())
 
     def __initByFile__(self, conf):
         self.changePort(conf)
@@ -62,11 +69,10 @@ class Port:
         self.__ip = option(self.__ip, conf['ip'])
         self.__mask = option(self.__mask, conf['mask'])
         self.__isUp = option(self.__ip, conf["isUp"])
-        try:
+        if self.__isActivate:
             self.__configPort__()
-        except:
-            logging.info('修改port配置失败.'+'newConf: '+str(conf)+self._get_msg())
-            # logging.info("[port: changePort]: 修改端口失败")
+
+
 
     def getID(self):
         return self.__id
@@ -85,6 +91,7 @@ class Port:
     def toJson(self):
         return {"id": self.__id,
                 "name": self.__name,
+                "isActive": self.__isActivate,
                 "regex": self.__pattern,
                 "ip": self.__ip,
                 "mask": self.__mask,
@@ -112,3 +119,5 @@ if __name__ == "__main__":
         m = pattern.match(s1)
         print(m.group(1))
         print(m.group(2))
+
+
